correção do controller do futuro cliente
